+++ test/common/ssl/utility_test.cc
@@ -3,6 +3,7 @@
 
 #include "common/ssl/utility.h"
 
+#include "test/common/ssl/test_utility.h"
 #include "test/test_common/environment.h"
 #include "test/test_common/tls_utility.h"
 #include "test/test_common/utility.h"
@@ -14,29 +15,26 @@ namespace Envoy {
 namespace Ssl {
 
 TEST(UtilityTest, TestGetSubjectAlternateNamesWithDNS) {
-  bssl::UniquePtr<X509> cert =
-      Envoy::Tls::Test::readCertFromFile("test/common/ssl/test_data/san_dns_cert.pem");
+  bssl::UniquePtr<X509> cert = readCertFromFile("test/common/ssl/test_data/san_dns_cert.pem");
   const std::vector<std::string>& subject_alt_names = Utility::getSubjectAltNames(*cert, GEN_DNS);
   EXPECT_EQ(1, subject_alt_names.size());
 }
 
 TEST(UtilityTest, TestMultipleGetSubjectAlternateNamesWithDNS) {
   bssl::UniquePtr<X509> cert =
-      Envoy::Tls::Test::readCertFromFile("test/common/ssl/test_data/san_multiple_dns_cert.pem");
+      readCertFromFile("test/common/ssl/test_data/san_multiple_dns_cert.pem");
   const std::vector<std::string>& subject_alt_names = Utility::getSubjectAltNames(*cert, GEN_DNS);
   EXPECT_EQ(2, subject_alt_names.size());
 }
 
 TEST(UtilityTest, TestGetSubjectAlternateNamesWithUri) {
-  bssl::UniquePtr<X509> cert =
-      Envoy::Tls::Test::readCertFromFile("test/common/ssl/test_data/san_uri_cert.pem");
+  bssl::UniquePtr<X509> cert = readCertFromFile("test/common/ssl/test_data/san_uri_cert.pem");
   const std::vector<std::string>& subject_alt_names = Utility::getSubjectAltNames(*cert, GEN_URI);
   EXPECT_EQ(1, subject_alt_names.size());
 }
 
 TEST(UtilityTest, TestGetSubjectAlternateNamesWithNoSAN) {
-  bssl::UniquePtr<X509> cert =
-      Envoy::Tls::Test::readCertFromFile("test/common/ssl/test_data/no_san_cert.pem");
+  bssl::UniquePtr<X509> cert = readCertFromFile("test/common/ssl/test_data/no_san_cert.pem");
   const std::vector<std::string>& uri_subject_alt_names =
       Utility::getSubjectAltNames(*cert, GEN_URI);
   EXPECT_EQ(0, uri_subject_alt_names.size());
