+++ test/test_common/tls_utility.h
@@ -3,10 +3,6 @@
 #include <string>
 #include <vector>
 
-#include "test/test_common/environment.h"
-
-#include "openssl/x509v3.h"
-
 namespace Envoy {
 namespace Tls {
 namespace Test {
@@ -19,13 +15,6 @@ namespace Test {
  *             in Application-Layer Protocol Negotiation. No ALPN is advertised if alpn is empty.
  */
 std::vector<uint8_t> generateClientHello(const std::string& sni_name, const std::string& alpn);
-
-/**
- * Reads a certificate from the given path.
- * @param path path of the certificate file to read from.
- * @return returns a pointer to the certificate.
- */
-bssl::UniquePtr<X509> readCertFromFile(const std::string& path);
 
 } // namespace Test
 } // namespace Tls
